// frontend/src/features/backoffice/administrativedistrcit/model/type.ts
export type AdministarativeDistrict = {
    type: string;
    code: number | string;
    level1: string;
    level2: string;
    level3: string;
    x: number | string;
    y: number | string;
    longitude_degrees: number | string;
    longitude_minutes: number | string;
    longitude_seconds: number | string;
    latitude_degrees: number | string;
    latitude_minutes: number | string;
    latitude_seconds: number | string;
    longitude: number | string;
    latitude: number | string;
    updatedAt: String
}